Use late escaping.
The algorithm previously did an escape-then-parse. Of course this was
somewhat wrong, but after trying to get it to work on Python 2, it was
noted that there was a minor difference in `re.escape`, namely the `_`
character was also escaped.

By moving the escaping to a later phase, the regular expressions become
a lot cleaner, and the algorithm also looks better. Escaping is only
done on the 'literal' parts, and not on the 'parameter' parts.